+++ Tikito/src/main/java/com/sunbeam/tikito/dto/EventResponseDto.java
@@ -18,7 +18,7 @@ public class EventResponseDto {
 
     private String eventDescription;
 
-    private Integer eventDurationMin;
+    private Long eventDurationMin;
 
     private Integer ageRestriction;
